+++ packages/fxa-content-server/app/tests/spec/views/settings/gravatar_permissions.js
@@ -9,6 +9,7 @@ define(function (require, exports, module) {
   const chai = require('chai');
   const Metrics = require('lib/metrics');
   const p = require('lib/promise');
+  const Notifier = require('lib/channels/notifier');
   const Relier = require('models/reliers/relier');
   const sinon = require('sinon');
   const TestHelpers = require('../../../lib/helpers');
@@ -55,6 +56,7 @@ define(function (require, exports, module) {
     function initView () {
       view = new View({
         metrics: metrics,
+        notifier: new Notifier(),
         relier: relier,
         user: user
       });